Make logging opt-in

+++ classes/Logging/MailTemplateLog.php
@@ -2,8 +2,24 @@
 
 namespace Stem\MailTemplates\Logging;
 
+use Stem\Core\Context;
+
 class MailTemplateLog {
+	private static $loggingEnabled = null;
+
+	private static function loadConfig() {
+		if (static::$loggingEnabled == null) {
+			static::$loggingEnabled = Context::current()->setting('mail-templates/log', false);
+		}
+	}
+
 	public static function LogSuccess($responseId, $response, $template, $to, $bcc, $data) {
+		static::loadConfig();
+
+		if (!static::$loggingEnabled) {
+			return;
+		}
+
 		/** @var \WPDB */
 		global $wpdb;
 		$wpdb->insert('stem_mail_template_log', [
@@ -18,6 +34,12 @@ class MailTemplateLog {
 	}
 
 	public static function LogError($template, $to, $bcc, $data, $error) {
+		static::loadConfig();
+
+		if (!static::$loggingEnabled) {
+			return;
+		}
+
 		/** @var \WPDB */
 		global $wpdb;
 		$wpdb->insert('stem_mail_template_log', [
@@ -31,6 +53,12 @@ class MailTemplateLog {
 	}
 
 	public static function Clear() {
+		static::loadConfig();
+
+		if (!static::$loggingEnabled) {
+			return;
+		}
+
 		global $wpdb;
 		$wpdb->query('delete from stem_mail_template_log');
 	}
